+++ src/modals/TaskModal.jsx
@@ -9,7 +9,8 @@ import AddEditTaskModal from "./AddEditTaskModal";
 
 const TaskModal = ({colIndex, taskIndex, setIsTaskModalOpen}) => {
     const dispatch = useDispatch();
-
+    const [isElipsisMenuOpen, setIsElipsisMenuOpen] = useState(false);
+    const [isDeleteModalOpen, setIsDeleteModalOpen] = useState(false);
     const boards = useSelector((state) => state.boards);
     const board = boards.find((board) => board.isActive === true);
     const columns = board.columns;
@@ -26,65 +27,68 @@ const TaskModal = ({colIndex, taskIndex, setIsTaskModalOpen}) => {
 
     const [status, setStatus] = useState(task.status);
     const [newColIndex, setNewColIndex] = useState(columns.indexOf(col));
-    const [elipsisMenuOpen, setElipsisMenuOpen] = useState(false);
-    const [isDeleteModalOpen, setIsDeleteModalOpen] = useState(false);
-    const [isAddTaskModalOpen, setIsAddTaskModalOpen] = useState(false);
-
-    const setOpenEditModal = () => {
-        setIsAddTaskModalOpen(true);
-        setElipsisMenuOpen(false);
-    };
-
-    const setOpenDeleteModal = () => {
-        setElipsisMenuOpen(false);
-        setIsDeleteModalOpen(true);
-    };
-
-    const onClose = (e) => {
-        if (e.target !== e.currentTarget) {
-            return
-        }
-
-        dispatch(
-            boardsSlice.actions.setTaskStatus({
-                taskIndex, colIndex, newColIndex, status
-            })
-        )
-        setIsTaskModalOpen(false);
-    };
-
     const onChange = (e) => {
         setStatus(e.target.value);
         setNewColIndex(e.target.selectedIndex);
     };
 
-    const onDeleteBtnClick = () => {
-        dispatch(boardsSlice.actions.deleteTask({taskIndex, colIndex}));
+    const onClose = (e) => {
+        if (e.target !== e.currentTarget) {
+            return;
+        }
+        dispatch(
+            boardsSlice.actions.setTaskStatus({
+                taskIndex,
+                colIndex,
+                newColIndex,
+                status,
+            })
+        );
         setIsTaskModalOpen(false);
-        setIsDeleteModalOpen(false);
+    };
+
+    const onDeleteBtnClick = (e) => {
+        if (e.target.textContent === "Delete") {
+            dispatch(boardsSlice.actions.deleteTask({ taskIndex, colIndex }));
+            setIsTaskModalOpen(false);
+            setIsDeleteModalOpen(false);
+        } else {
+            setIsDeleteModalOpen(false);
+        }
+    };
+
+    const [isAddTaskModalOpen, setIsAddTaskModalOpen] = useState(false);
+
+    const setOpenEditModal = () => {
+        setIsAddTaskModalOpen(true);
+        setIsElipsisMenuOpen(false);
+    };
+
+    const setOpenDeleteModal = () => {
+        setIsElipsisMenuOpen(false);
+        setIsDeleteModalOpen(true);
     };
 
     return (
-        <div onClick={onClose}
-             className='fixed right-0 left-0 top-0 py-4 overflow-scroll scrollbar-hide z-50 bottom-0 justify-center
-        items-center flex bg-[#00000080]'>
+        <div
+            onClick={onClose}
+            className=" fixed right-0 top-0 px-2 py-4 overflow-scroll scrollbar-hide  z-50 left-0 bottom-0 justify-center items-center flex dropdown "
+        >
+            {/* MODAL SECTION */}
 
-            {/*Modal Section*/}
-            <div
-
-                className=" scrollbar-hide overflow-y-scroll max-h-[95vh]  my-auto  bg-white dark:bg-[#2b2c37] text-black dark:text-white font-bold shadow-md shadow-[#364e7e1a] max-w-md mx-auto  w-full px-8  py-8 rounded-xl">
+            <div className=" scrollbar-hide overflow-y-scroll max-h-[95vh]  my-auto  bg-white dark:bg-[#2b2c37] text-black dark:text-white font-bold shadow-md shadow-[#364e7e1a] max-w-md mx-auto  w-full px-8  py-8 rounded-xl">
                 <div className=" relative flex   justify-between w-full items-center">
                     <h1 className=" text-lg">{task.title}</h1>
 
                     <img
                         onClick={() => {
-                            setElipsisMenuOpen((state) => !state);
+                            setIsElipsisMenuOpen((prevState) => !prevState);
                         }}
                         src={elipsis}
                         alt="elipsis"
                         className=" cursor-pointer h-6"
                     />
-                    {elipsisMenuOpen && (
+                    {isElipsisMenuOpen && (
                         <ElipsisMenu
                             setOpenEditModal={setOpenEditModal}
                             setOpenDeleteModal={setOpenDeleteModal}
@@ -100,46 +104,59 @@ const TaskModal = ({colIndex, taskIndex, setIsTaskModalOpen}) => {
                     Subtasks ({completed} of {subtasks.length})
                 </p>
 
-                {/*    Subtask section*/}
-                <div className='mt-3 space-y-2'>
-                    {subtasks.map((subtask, i) => {
+                {/* subtasks section */}
+
+                <div className=" mt-3 space-y-2">
+                    {subtasks.map((subtask, index) => {
                         return (
-                            <Subtask index={i} taskIndex={taskIndex} colIndex={colIndex}
-                                     key={i}/>
-                        )
+                            <Subtask
+                                index={index}
+                                taskIndex={taskIndex}
+                                colIndex={colIndex}
+                                key={index}
+                            />
+                        );
                     })}
                 </div>
 
-                {/*Current Status Section*/}
-                <div className='mt-8 flex flex-col space-y-3'>
-                    <label className='text-sm dark:text-white text-gray-500'>
+                {/* Current Status Section */}
+
+                <div className="mt-8 flex flex-col space-y-3">
+                    <label className="  text-sm dark:text-white text-gray-500">
                         Current Status
                     </label>
-
-                    <select className='select-status flex-grow px-4 py-2 rounded-md
-                    text-sm bg-transparent focus:border-0 border border-gray-300 focus:outline-[#635fc7]
-                    outline-none' value={status} onChange={onChange}>
-                        {columns.map((column, index) => (
-                            <option className='status-option' key={index}>
-                                {column.name}
+                    <select
+                        className=" select-status flex-grow px-4 py-2 rounded-md text-sm bg-transparent focus:border-0  border-[1px] border-gray-300 focus:outline-[#635fc7] outline-none"
+                        value={status}
+                        onChange={onChange}
+                    >
+                        {columns.map((col, index) => (
+                            <option className="status-options" key={index}>
+                                {col.name}
                             </option>
                         ))}
                     </select>
                 </div>
             </div>
-
             {isDeleteModalOpen && (
                 <DeleteModal
-                    setIsDeleteModalOpen={setIsDeleteModalOpen}
-                    onDeleteBtnClick={onDeleteBtnClick} title={task.title} type='task'/>
+                    onDeleteBtnClick={onDeleteBtnClick}
+                    type="task"
+                    title={task.title}
+                />
             )}
 
             {isAddTaskModalOpen && (
-                <AddEditTaskModal setOpenAddEditTask={setIsAddTaskModalOpen} type='edit' taskIndex={taskIndex}
-                                  prevColIndex={colIndex} setIsTaskModalOpen={setIsTaskModalOpen}/>
+                <AddEditTaskModal
+                    setIsAddTaskModalOpen={setIsAddTaskModalOpen}
+                    setIsTaskModalOpen={setIsTaskModalOpen}
+                    type="edit"
+                    taskIndex={taskIndex}
+                    prevColIndex={colIndex}
+                />
             )}
         </div>
-    )
-};
+    );
+}
 
 export default TaskModal;
